Move pattern note commands to a separate file
