+++ lib/widgets/editors/piano_roll/controller/piano_roll_controller.dart
@@ -20,7 +20,7 @@
 import 'dart:math';
 
 import 'package:anthem/commands/journal_commands.dart';
-import 'package:anthem/commands/pattern_commands.dart';
+import 'package:anthem/commands/pattern_note_commands.dart';
 import 'package:anthem/commands/timeline_commands.dart';
 import 'package:anthem/helpers/id.dart';
 import 'package:anthem/model/pattern/note.dart';
